Do not move temp files but copy them.
This avoids issues on Linux when /tmp is on a different filesystem than NOTION_HOME. In the long run it's probably a better idea to make sure all temp files are created in NOTION_HOME/tmp (?) imho, so persisting the files can again be an atomic move. For now this should help a lot of people on Linux.